Fix the visibility changed again
